lawgpt-backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-firestore
uvicorn[standard]
python-dotenv
pdfplumber
